AST-18457 - CLI | SCA Real-time | Create new command - Handle SCA Rest APIs
- Call Sca Apis to handle sca realtime
