+++ internal/commands/scarealtime/sca-linux-mac-utils.go
@@ -1,0 +1,73 @@
+//go:build linux || darwin
+
+package scarealtime
+
+import (
+	"archive/tar"
+	"compress/gzip"
+	"fmt"
+	"github.com/checkmarx/ast-cli/internal/logger"
+	"io"
+	"io/fs"
+	"log"
+	"os"
+	"path/filepath"
+)
+
+// UnzipOrExtractFiles Extracts SCA Resolver files
+func UnzipOrExtractFiles() error {
+	logger.PrintIfVerbose("Extracting files in: " + ScaResolverWorkingDir)
+	gzipStream, err := os.Open(filepath.Join(ScaResolverWorkingDir, Params.SCAResolverFileName))
+	if err != nil {
+		fmt.Println("error")
+	}
+	uncompressedStream, err := gzip.NewReader(gzipStream)
+	if err != nil {
+		log.Fatal("ExtractTarGz: NewReader failed")
+	}
+
+	tarReader := tar.NewReader(uncompressedStream)
+
+	for {
+		header, err := tarReader.Next()
+
+		if err == io.EOF {
+			break
+		}
+
+		if err != nil {
+			log.Fatalf("ExtractTarGz: Next() failed: %s", err.Error())
+		}
+
+		switch header.Typeflag {
+		case tar.TypeDir:
+			if err := os.Mkdir(header.Name, 0755); err != nil { //nolint:gomnd
+				log.Fatalf("ExtractTarGz: Mkdir() failed: %s", err.Error())
+			}
+		case tar.TypeReg:
+			extractedFilePath := filepath.Join(ScaResolverWorkingDir, header.Name)
+			outFile, err := os.Create(extractedFilePath)
+			if err != nil {
+				log.Fatalf("ExtractTarGz: Create() failed: %s", err.Error())
+			}
+			if _, err = io.Copy(outFile, tarReader); err != nil {
+				log.Fatalf("ExtractTarGz: Copy() failed: %s", err.Error())
+			}
+			err = outFile.Close()
+			if err != nil {
+				return err
+			}
+			err = os.Chmod(extractedFilePath, fs.ModePerm)
+			if err != nil {
+				return err
+			}
+		default:
+			log.Fatalf(
+				"ExtractTarGz: uknown type: %v in %s",
+				header.Typeflag,
+				header.Name)
+		}
+	}
+
+	return nil
+}
